replace paragraph tag with div tag for note text
note text was placed in a paragraph tag in the show template; this fails to generate valid HTML because it nests paragraph tags if the note text contains them (since the note text often contains HTML).  More importantly, if the generated HTML is parsed/the content is extracted by id, it will only extract the first paragraph portion (not the remainder).